Remplaced default user_field by field according the current platform
